@@ -17,6 +17,7 @@
     "@angular/common": "^6.1.0",
     "@angular/compiler": "^6.1.0",
     "@angular/core": "^6.1.0",
+    "@angular/flex-layout": "^6.0.0-beta.18",
     "@angular/forms": "^6.1.0",
     "@angular/http": "^6.1.0",
     "@angular/material": "^6.4.7",
@@ -32,6 +33,7 @@
     "core-js": "^2.5.4",
     "hammerjs": "^2.0.8",
     "moment": "^2.22.2",
+    "ngx-facebook": "^2.4.0",
     "rxjs": "^6.0.0",
     "zone.js": "~0.8.26"
   },
@@ -54,6 +56,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "^2.9.2"
   }
 }
