@@ -4,6 +4,7 @@ import {first, tap} from 'rxjs/operators';
 import {Router} from '@angular/router';
 import {ApiService} from '../shared/services/api.service';
 import {GrantType} from './types/grant-type';
+import {FacebookService, InitParams, LoginOptions, LoginResponse} from 'ngx-facebook';
 
 declare const gapi: any;
 
@@ -15,7 +16,8 @@ export class AuthService {
   constructor(
     private apiService: ApiService,
     private router: Router,
-    private ngZone: NgZone
+    private ngZone: NgZone,
+    private fb: FacebookService
   ) {
     this.googleInit();
   }
@@ -47,17 +49,50 @@ export class AuthService {
         backend: 'google-oauth2',
         token: tokenId
       };
-      return this.apiService.post('auth/convert-token', body)
-        .pipe(first())
-        .toPromise()
-        .then((credentials: any) => this.apiService.saveToken(credentials))
-        .then(() => {
-          this.ngZone.run(() => {
-              this.router.navigate(['app', 'news']);
-            }
-          );
-        });
+      return this.convertToken(body);
     });
+  }
+
+
+  async loginWithFacebook() {
+
+    // const options: LoginOptions = {
+    //   scope: 'public_profile,email',
+    //   return_scopes: true,
+    //   enable_profile_selector: true
+    // };
+    const initParams: InitParams = {
+      appId: 'OJPMVO5wbbvY6FwW2Cb922IPgeJf3DOR0Xr5rUyn',
+      xfbml: true,
+      version: 'v2.8',
+    };
+
+    await this.fb.init(initParams)
+      .then(() => this.fb.login())
+      .then((loginResponse: LoginResponse) => {
+        const body = {
+          client_id: environment.clientId,
+          client_secret: environment.facebookClientSecret,
+          grant_type: GrantType.CONVERT_TOKEN,
+          backend: 'facebook',
+          token: loginResponse.authResponse.accessToken
+        };
+        return this.convertToken(body);
+      })
+      .catch((error: any) => console.error(error));
+  }
+
+  convertToken(body) {
+    this.apiService.post('auth/convert-token', body)
+      .pipe(first())
+      .toPromise()
+      .then((credentials: any) => this.apiService.saveToken(credentials))
+      .then(() => {
+        this.ngZone.run(() => {
+            this.router.navigate(['app', 'news']);
+          }
+        );
+      });
   }
 
   logout() {
