@@ -5,8 +5,7 @@ import {Observable} from 'rxjs';
 import {select, Store} from '@ngrx/store';
 import * as newsStore from '../../store';
 import * as fromRoot from '../../../store';
-import {MatDialog} from '@angular/material';
-import {NewsPageComponent} from '../news-page/news-page.component';
+
 
 @Component({
   selector: 'news-feed',
@@ -19,7 +18,8 @@ import {NewsPageComponent} from '../news-page/news-page.component';
         [audio]="audio$ | async"
         (pause)="onPause($event)"
         (play)="onPlay($event)"
-        (view)="onView($event)">
+        (view)="onView($event)"
+        (openUrl)="onOpenUrl($event)">
       </news-list>
   `,
   styleUrls: ['./news-feed.component.scss'],
@@ -34,7 +34,7 @@ export class NewsFeedComponent implements OnInit {
 
   constructor(
     private newsService: NewsService,
-    private store: Store<newsStore.NewsState>,
+    private store: Store<newsStore.NewsState>
   ) { }
 
   ngOnInit() {
@@ -57,5 +57,9 @@ export class NewsFeedComponent implements OnInit {
     this.store.dispatch(new fromRoot.Go({
       path: ['/', 'app', 'news', event.id]
     }));
+  }
+
+  onOpenUrl(event: {url: string, config: any}) {
+    window.open(event.url, '_blank');
   }
 }
