@@ -7,6 +7,7 @@ import { NotFoundComponent } from './components/not-found/not-found.component';
 import {ScrollableDirective} from './directives/scrollable.directive';
 import { TextCopyDirective } from './directives/copy-url.directive';
 import { MinuteSecondsPipe } from './pipes/minute-seconds.pipe';
+import { ShareComponent } from './components/share/share.component';
 
 @NgModule({
   imports: [
@@ -16,7 +17,7 @@ import { MinuteSecondsPipe } from './pipes/minute-seconds.pipe';
     AppMaterialModule,
     HttpClientModule,
   ],
-  declarations: [NotFoundComponent, ScrollableDirective, TextCopyDirective, MinuteSecondsPipe],
+  declarations: [NotFoundComponent, ScrollableDirective, TextCopyDirective, MinuteSecondsPipe, ShareComponent],
   exports: [
     FormsModule,
     ReactiveFormsModule,
@@ -25,7 +26,8 @@ import { MinuteSecondsPipe } from './pipes/minute-seconds.pipe';
     NotFoundComponent,
     ScrollableDirective,
     TextCopyDirective,
-    MinuteSecondsPipe
+    MinuteSecondsPipe,
+    ShareComponent
   ]
 })
 export class SharedModule { }
